style: 메인페이지 nav바 스타일 수정
