+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Inter as FontSans } from "next/font/google";
 import { cn } from "@/lib/utils";
 import { isThereServerSession } from "@/hooks/session";
-import NavBar from "../components/customUi/NavBar";
+import NavBar from "../components/customUi/navUi/NavBar";
 import "./globals.css";
 import QueryProvider from "./provider";
 import Link from "next/link";
@@ -33,17 +33,15 @@ export default async function RootLayout({
             "bg-lime-400"
           )}
         >
+          <NavBar />
           <div className="w-[1500px] mx-auto">
-            <nav>
-              <NavBar />
-            </nav>
-            <Link href="/" className="m-2">
+            {/* <Link href="/" className="m-2">
               로고
             </Link>
             <Link href="/community" className="m-2">
               커뮤니티
             </Link>
-            <Button>로그아웃</Button>
+            <Button>로그아웃</Button> */}
             {children}
           </div>
         </body>
