--- src/components/customUi/NavBar.tsx
@@ -1,38 +1,0 @@
-"use client";
-import { useRouter } from "next/navigation";
-import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
-import { zustandStore } from "@/store/zustandStore";
-import { useEffect } from "react";
-
-const NavBar = () => {
-  const router = useRouter();
-  const supabase = createClientComponentClient();
-  const loggedIn = zustandStore((state) => state.loggedIn);
-  const changeLoggedIn = zustandStore((state) => state.changeLoggedIn);
-
-  useEffect(() => {
-    const fetchSession = async () => {
-      const { data } = await supabase.auth.getSession();
-      changeLoggedIn(!!data.session);
-    };
-    fetchSession();
-  }, [supabase.auth, changeLoggedIn]);
-
-  const handleSignOut = async () => {
-    await supabase.auth.signOut();
-    changeLoggedIn(false);
-    router.refresh();
-  };
-
-  const goToLoginpage = () => {
-    router.push("/login");
-  };
-
-  return loggedIn ? (
-    <button onClick={handleSignOut}>로그아웃</button>
-  ) : (
-    <button onClick={goToLoginpage}>로그인</button>
-  );
-};
-
-export default NavBar;
